can: Merge netpacket/can.h into nuttx/can.h
To align with the layout of Linux can header file.
