+++ arch/arm/src/s32k1xx/s32k1xx_flexcan.c
@@ -33,7 +33,6 @@
 #include <debug.h>
 #include <errno.h>
 
-#include <nuttx/can.h>
 #include <nuttx/wdog.h>
 #include <nuttx/irq.h>
 #include <nuttx/arch.h>
