+++ arch/arm/src/stm32/stm32_fdcan_sock.c
@@ -38,10 +38,8 @@
 #include <nuttx/arch.h>
 
 #include <nuttx/wqueue.h>
-#include <nuttx/can.h>
 #include <nuttx/net/netdev.h>
 #include <nuttx/net/can.h>
-#include <netpacket/can.h>
 
 #include "arm_internal.h"
 #include "stm32_fdcan.h"
