+++ boards/arm/s32k3xx/mr-canhubk3/src/s32k3xx_tja1153.c
@@ -39,7 +39,6 @@
 #include <sys/ioctl.h>
 #include <sys/socket.h>
 #include <nuttx/can.h>
-#include <netpacket/can.h>
 #include <nuttx/signal.h>
 
 #include "s32k3xx_pin.h"
